+++ src/Curate/Structured/GpsMetadata.php
@@ -34,7 +34,7 @@ final readonly class GpsMetadata
             return null;
         }
 
-        return new GpsCoordinate($this->gps->latitude, $this->gps->latitudeRef);
+        return new GpsCoordinate($this->gps->latitude, $this->gps->latitudeReference());
     }
 
     public function longitude(): ?GpsCoordinate
@@ -43,7 +43,7 @@ final readonly class GpsMetadata
             return null;
         }
 
-        return new GpsCoordinate($this->gps->longitude, $this->gps->longitudeRef);
+        return new GpsCoordinate($this->gps->longitude, $this->gps->longitudeReference());
     }
 
     public function altitude(): ?float
@@ -53,7 +53,7 @@ final readonly class GpsMetadata
 
     public function altitudeReference(): ?int
     {
-        return $this->gps->altitudeRef;
+        return $this->gps->altitudeReference();
     }
 
     public function version(): ?string
@@ -83,12 +83,12 @@ final readonly class GpsMetadata
 
     public function dilutionOfPrecision(): ?float
     {
-        return $this->gps->dop;
+        return $this->gps->dilutionOfPrecision();
     }
 
     public function speedReference(): ?string
     {
-        return $this->gps->speedRef;
+        return $this->gps->speedReference();
     }
 
     public function speedMs(): ?float
@@ -98,7 +98,7 @@ final readonly class GpsMetadata
 
     public function speedOriginalReference(): ?string
     {
-        return $this->gps->speedOriginalRef;
+        return $this->gps->speedOriginalReference();
     }
 
     public function speedOriginal(): ?float
@@ -108,7 +108,7 @@ final readonly class GpsMetadata
 
     public function trackReference(): ?string
     {
-        return $this->gps->trackRef;
+        return $this->gps->trackReference();
     }
 
     public function track(): ?float
@@ -118,7 +118,7 @@ final readonly class GpsMetadata
 
     public function imageDirectionReference(): ?string
     {
-        return $this->gps->imageDirectionRef;
+        return $this->gps->imageDirectionReference();
     }
 
     public function imageDirection(): ?float
@@ -137,7 +137,7 @@ final readonly class GpsMetadata
             return null;
         }
 
-        return new GpsCoordinate($this->gps->destinationLatitude, $this->gps->destinationLatitudeRef);
+        return new GpsCoordinate($this->gps->destinationLatitude, $this->gps->destinationLatitudeReference());
     }
 
     public function destinationLongitude(): ?GpsCoordinate
@@ -146,12 +146,12 @@ final readonly class GpsMetadata
             return null;
         }
 
-        return new GpsCoordinate($this->gps->destinationLongitude, $this->gps->destinationLongitudeRef);
+        return new GpsCoordinate($this->gps->destinationLongitude, $this->gps->destinationLongitudeReference());
     }
 
     public function destinationBearingReference(): ?string
     {
-        return $this->gps->destinationBearingRef;
+        return $this->gps->destinationBearingReference();
     }
 
     public function destinationBearing(): ?float
@@ -161,7 +161,7 @@ final readonly class GpsMetadata
 
     public function destinationDistanceReference(): ?string
     {
-        return $this->gps->destinationDistanceRef;
+        return $this->gps->destinationDistanceReference();
     }
 
     public function destinationDistanceMetres(): ?float
@@ -171,7 +171,7 @@ final readonly class GpsMetadata
 
     public function destinationDistanceOriginalReference(): ?string
     {
-        return $this->gps->destinationDistanceOriginalRef;
+        return $this->gps->destinationDistanceOriginalReference();
     }
 
     public function destinationDistanceOriginal(): ?float